+++ misc/run_tests.py
@@ -1163,10 +1163,8 @@ async def main():
 
             scenes = [
                 "data/picort/barbarian.picort.txt",
-                "data/picort/barbarian-big.picort.txt",
                 "data/picort/slime-binary.picort.txt",
-                "data/picort/slime-ascii.picort.txt",
-                "data/picort/slime-big.picort.txt",
+                "data/picort/material-chart.picort.txt",
             ]
 
             for scene in scenes:
